YouTube Assistant/requirements.txt
langchain
openai
streamlit
python-dotenv